+++ NEWS.md
@@ -1,5 +1,7 @@
 # dtplyr (development version)
 
+* `T` and `F` are correctly translated to `TRUE` and `FALSE` (#140).
+
 # dtplyr 1.0.1
 
 * Better handling for `.data` and `.env` pronouns (#138).